Move GetWarnings() into its own file.
