+++ src/init.cpp
@@ -44,6 +44,7 @@
 #include "wallet/wallet.h"
 #include "wallet/walletdb.h"
 #endif
+#include "warnings.h"
 #include <stdint.h>
 #include <stdio.h>
